reddit_scrapper/requirements.txt

beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
openpyxl==3.1.2
numpy==1.26.4
pandas==2.2.2
selenium==4.19.0
webdriver-manager==4.0.1